+++ src/com/appriskgame/controller/Player.java
@@ -11,10 +11,9 @@ import java.util.Random;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import com.appriskgame.model.Card;
+import com.appriskgame.model.CardController;
 import com.appriskgame.model.Continent;
 import com.appriskgame.model.Country;
-import com.appriskgame.model.Deck;
 import com.appriskgame.model.GameMap;
 import com.appriskgame.model.GamePlayer;
 
@@ -52,105 +51,6 @@ public class Player {
 	public static List<String> playersChoiceList = new ArrayList<String>();
 
 	public boolean doFortification = false;
-	private int NumOfTypeCardsExchanged = 0;
-	private String SameCardExchangeThrice = "";
-
-	private Deck deck = Deck.getInstance();
-	private boolean exchanged = false;
-	public int playerCardExchangeCount = 0;
-	private List<Card> cardList = new ArrayList<>();
-
-	/**
-	 * This method returns the list of cards player has
-	 *
-	 * @return cardList list of cards
-	 */
-	public List<Card> getCardList() {
-		return cardList;
-	}
-
-	/**
-	 * This method sets the list of cards player has
-	 *
-	 * @return cardList list of cards
-	 */
-	public void setCardList(List<Card> cardList) {
-		this.cardList = cardList;
-	}
-
-	/**
-	 * This method gives the count for player's card exchange.
-	 *
-	 * @return
-	 */
-	public int getPlayerCardExchangeCount() {
-		return playerCardExchangeCount;
-	}
-
-	/**
-	 * This method sets the count for player's card exchange.
-	 *
-	 * @param playerCardExchangeCount
-	 */
-	public void setPlayerCardExchangeCount(int playerCardExchangeCount) {
-		this.playerCardExchangeCount = playerCardExchangeCount;
-	}
-
-	/**
-	 * This method does the exchange.
-	 *
-	 * @return
-	 */
-	public boolean isExchanged() {
-		return exchanged;
-	}
-
-	/**
-	 * This method sets the exchanged ones.
-	 *
-	 * @param exchanged
-	 */
-	public void setExchanged(boolean exchanged) {
-		this.exchanged = exchanged;
-	}
-
-	/**
-	 * This method returns the number of type of cards
-	 *
-	 * @return NumOfTypeCardsExchanged number of type of cards
-	 */
-	public int getNumOfTypeCardsExchanged() {
-		return NumOfTypeCardsExchanged;
-	}
-
-	/**
-	 * This method sets the number of type of cards
-	 *
-	 * @param numOfTypeCardsExchanged number of type of cards
-	 */
-	public void setNumOfTypeCardsExchanged(int numOfTypeCardsExchanged) {
-		NumOfTypeCardsExchanged = numOfTypeCardsExchanged;
-	}
-
-	/**
-	 * This method returns card which appear more than thrice during exchange
-	 *
-	 * @return SameCardExchangeThrice card which appear more that thrice in exchange
-	 */
-	public String getSameCardExchangeThrice() {
-		return SameCardExchangeThrice;
-	}
-
-	/**
-	 * This method sets the number of cards which appear more than thrice during
-	 * exchange
-	 *
-	 * @param sameCardExchangeThrice card which appear more that thrice in exchange
-	 */
-	public void setSameCardExchangeThrice(String sameCardExchangeThrice) {
-		SameCardExchangeThrice = sameCardExchangeThrice;
-	}
-
 	/**
 	 * This method is used to set the player's list.
 	 *
@@ -363,7 +263,7 @@ public class Player {
 	// startup phase methods
 	/**
 	 * This Method is used to Add and remove players to playerList.
-	 * 
+	 *
 	 * @return List of Player Names
 	 */
 	public ArrayList<String> playerCreation() {
@@ -626,8 +526,10 @@ public class Player {
 				listOfContries.add(country);
 			}
 		}
+		CardController cardController = new CardController();
+		int exchangeInCard = cardController.exchangeCards(player);
 		int reinforcementArmies = assignReinforcedArmies(player, playerContinent);
-		player.setNoOfArmies((reinforcementArmies));
+		player.setNoOfArmies((reinforcementArmies + exchangeInCard));
 
 		while (player.getNoOfArmies() > 0) {
 			System.out.println(" Player Name :" + player.getPlayerName());
@@ -1710,118 +1612,5 @@ public class Player {
 			System.out.println("Countries are not adjacent!");
 			doFortification = true;
 		}
-	}
-
-	/**
-	 * Checks if the given three types of cards are same
-	 *
-	 * @param exchangeCards       exchange cards
-	 * @param cardAppearingThrice Card's name which is appearing thrice
-	 * @return true if all are same else false
-	 */
-	public boolean checkCardSameType(List<Card> exchangeCards, String cardAppearingThrice) {
-		int cardAppearingCount = 0;
-		if (exchangeCards.size() < 3) {
-			return false;
-		} else {
-			for (Card card : exchangeCards) {
-				if (card.getType().equals(cardAppearingThrice)) {
-					cardAppearingCount++;
-				}
-			}
-			if (cardAppearingCount == 3) {
-				SameCardExchangeThrice = cardAppearingThrice;
-				return true;
-			}
-		}
-		return false;
-	}
-
-	/**
-	 * To check exchanging cards are of different types
-	 *
-	 * @param exchangeCards exchange cards
-	 * @param cardTypes     card types
-	 * @return true or false depending on the types of cards
-	 */
-	public boolean checkDiffTypesOfCards(List<Card> exchangeCards, int cardTypes) {
-		if (cardTypes < 3) {
-			return false;
-		}
-		List<String> types = new ArrayList<>();
-		for (Card card : exchangeCards) {
-			types.add(card.getType());
-		}
-		if (!types.get(0).equals(types.get(1)) && !types.get(1).equals(types.get(2))
-				&& !types.get(2).equals(types.get(0))) {
-			NumOfTypeCardsExchanged = 3;
-			return true;
-		}
-		return false;
-	}
-
-	/**
-	 * This method helps to get the cards from the player's list and add to the deck
-	 * of cards
-	 *
-	 * @param cardTypes                   Type of cards
-	 * @param cardAppearingMoreThanThrice Card name appearing thrice
-	 * @param player                      Game player object
-	 * @param cardNumbers                 the cardNumbers
-	 * @throws Exception exception of method
-	 */
-	public void exchangeCards(int cardTypes, String cardAppearingMoreThanThrice, GamePlayer player,
-			List<Integer> cardNumbers) throws Exception {
-
-		if (cardTypes == 3 || (cardAppearingMoreThanThrice != null && !cardAppearingMoreThanThrice.isEmpty())) {
-			Card card1 = getCardList().get(cardNumbers.get(0) - 1);
-			Card card2 = getCardList().get(cardNumbers.get(1) - 1);
-			Card card3 = getCardList().get(cardNumbers.get(2) - 1);
-			getCardList().remove(card1);
-			getCardList().remove(card2);
-			getCardList().remove(card3);
-
-			deck.add(card1);
-			deck.add(card2);
-			deck.add(card3);
-			setExchanged(true);
-			setCardList(getCardList());
-//			setChanged();
-//			notifyObservers(player);
-		}
-	}
-
-	/**
-	 * If player has the card territory, he will get the extra set of armies
-	 *
-	 *
-	 * @param exchangeCards list of cards
-	 * @param player        GamePlayer object
-	 * @return playerObject GamePlayer Object
-	 */
-	public GamePlayer exChangeCardTerritoryExist(List<Card> exchangeCards, GamePlayer player) {
-		GamePlayer playerObject = new GamePlayer();
-		playerObject = player;
-		ArrayList<Country> countryList = new ArrayList<Country>();
-		countryList = player.getPlayerCountries();
-		ArrayList<Country> updatedCountryList = new ArrayList<Country>();
-		System.out.println("Inside Exchange card Territory");
-		for (Country country : countryList) {
-			for (Card card : exchangeCards) {
-				String[] CountryNameInCard = card.getName().split(",");
-				String countryName = CountryNameInCard[0];
-				if (countryName.equalsIgnoreCase(country.getCountryName())) {
-					System.out.println("country name for the exchnage is :" + country.getCountryName());
-					System.out.println("armies before exchnage in the Country is :" + country.getNoOfArmies());
-					int NumOfarmies = country.getNoOfArmies();
-					NumOfarmies = NumOfarmies + 2;
-					country.setNoOfArmies(NumOfarmies);
-					System.out.println("armies before exchnage in the Country is :" + country.getNoOfArmies());
-				}
-			}
-			updatedCountryList.add(country);
-		}
-		playerObject.setPlayerCountries(updatedCountryList);
-		return playerObject;
 	}
 }
