+++ src/com/appriskgame/model/Deck.java
@@ -1,17 +1,15 @@
 package com.appriskgame.model;
 
-import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.Collections;
-import java.util.List;
 
 /**
- * This class allows the creation of deck of risk cards
+ * This class allows the creation of deck of 42 risk cards
  *
  * @author Shruthi
  *
  */
-public class Deck implements Serializable {
+public class Deck {
 
 	private static Deck deck;
 
@@ -22,28 +20,39 @@ public class Deck implements Serializable {
 	/**
 	 * Creates all cards, one for each territory. Each card has either a type of
 	 * Infantry, Cavalry, or Artillery.
-	 * 
+	 *
 	 * @param list list of country names
 	 */
-	public void setDeckOfCards(List<Country> listOfCountries) {
+	public void setDeckOfCards() {
 
-		Collections.shuffle(listOfCountries);
 
 		typeOfCards = new String[] { "Infantry", "Cavalry", "Artillery" };
 
 		deckOfCards = new ArrayList<Card>();
 
-		for (int i = 0; i < listOfCountries.size(); i++) {
-			// Adding new instance of cards to deck
-			deckOfCards.add(new Card(typeOfCards[i % 3], listOfCountries.get(i)));
-			// System.out.println("Added new card to deck: " + deck.get(i).getName());
+		for (int i = 1; i <= 14; i++) {
+			Card card = new Card();
+			card.settype(typeOfCards[0]);
+			deckOfCards.add(card);
+		}
+		for (int i = 1; i <= 14; i++) {
+
+			Card card = new Card();
+			card.settype(typeOfCards[1]);
+			deckOfCards.add(card);
+		}
+		for (int i = 1; i <= 14; i++) {
+
+			Card card = new Card();
+			card.settype(typeOfCards[2]);
+			deckOfCards.add(card);
 		}
 		Collections.shuffle(deckOfCards);
 	}
 
 	/**
 	 * To get the instance of the Deck
-	 * 
+	 *
 	 * @return It is returning of type deck
 	 */
 	public static Deck getInstance() {
@@ -55,7 +64,7 @@ public class Deck implements Serializable {
 
 	/**
 	 * Removing a card from the deck and returning it
-	 * 
+	 *
 	 * @return card object
 	 */
 	public Card draw() {
@@ -68,7 +77,7 @@ public class Deck implements Serializable {
 
 	/**
 	 * This method adds a card to the deck
-	 * 
+	 *
 	 * @param card object of the card which is to be added to deck
 	 */
 	public void add(Card card) {
